refactor: rename bootstrap-test-application to application-testing
