+++ packages/spectacular/src/lib/application-testing/util-dom/bootstrap-component.ts
