+++ packages/spectacular/src/lib/application-testing/util-dom/ensure-root-element.ts
