+++ packages/spectacular/src/lib/application-testing/util-dom/get-content.ts
