+++ packages/spectacular/src/lib/application-testing/util-dom/is-template-element.ts
