+++ packages/spectacular/src/lib/application-testing/util-dom/replace-root-element.ts
